implemented more requirements
Front-end (client):
- display internet connection status
- display server connection status
- period check that data on client is synced with data from server
- added cap to pie chart legend
- handle connection errors

Back-end (server):
- added a fake data generator that runs on a separate thread

+++ backend/src/main/java/mpp/backend/Model/ComputerComponent.java
@@ -18,7 +18,7 @@ public class ComputerComponent {
     private String realid;
 
     @Indexed(unique = true)
-    private int productID;
+    private long productID;
     private String manufacturer;
     private String category;
     private String productName;
@@ -26,7 +26,7 @@ public class ComputerComponent {
     private Date releaseDate;
     private Integer quantity;
 
-    public ComputerComponent(int productID, String manufacturer, String productName, String category, double price, Integer quantity, Date releaseDate) {
+    public ComputerComponent(long productID, String manufacturer, String productName, String category, double price, Integer quantity, Date releaseDate) {
         this.productID = productID;
         this.manufacturer = manufacturer;
         this.productName = productName;
